Remove TaskInterface
