+++ src/ray/core_worker/core_worker.cc
@@ -1,7 +1,43 @@
 #include <boost/asio/signal_set.hpp>
 
+#include "ray/common/task/task_util.h"
 #include "ray/core_worker/context.h"
 #include "ray/core_worker/core_worker.h"
+
+namespace {
+
+void BuildCommonTaskSpec(
+    ray::TaskSpecBuilder &builder, const JobID &job_id, const TaskID &task_id,
+    const TaskID &current_task_id, const int task_index, const TaskID &caller_id,
+    const ray::RayFunction &function, const std::vector<ray::TaskArg> &args,
+    uint64_t num_returns,
+    const std::unordered_map<std::string, double> &required_resources,
+    const std::unordered_map<std::string, double> &required_placement_resources,
+    ray::TaskTransportType transport_type, std::vector<ObjectID> *return_ids) {
+  // Build common task spec.
+  builder.SetCommonTaskSpec(task_id, function.GetLanguage(),
+                            function.GetFunctionDescriptor(), job_id, current_task_id,
+                            task_index, caller_id, num_returns, required_resources,
+                            required_placement_resources);
+  // Set task arguments.
+  for (const auto &arg : args) {
+    if (arg.IsPassedByReference()) {
+      builder.AddByRefArg(arg.GetReference());
+    } else {
+      builder.AddByValueArg(arg.GetValue());
+    }
+  }
+
+  // Compute return IDs.
+  (*return_ids).resize(num_returns);
+  for (size_t i = 0; i < num_returns; i++) {
+    (*return_ids)[i] =
+        ObjectID::ForTaskReturn(task_id, i + 1,
+                                /*transport_type=*/static_cast<int>(transport_type));
+  }
+}
+
+}  // namespace
 
 namespace ray {
 
@@ -48,8 +84,6 @@ CoreWorker::CoreWorker(
   object_interface_ =
       std::unique_ptr<CoreWorkerObjectInterface>(new CoreWorkerObjectInterface(
           worker_context_, raylet_client_, store_socket, use_memory_store));
-  task_interface_ = std::unique_ptr<CoreWorkerTaskInterface>(new CoreWorkerTaskInterface(
-      worker_context_, raylet_client_, *object_interface_, io_service_));
 
   // Initialize task execution.
   int rpc_server_port = 0;
@@ -97,6 +131,11 @@ CoreWorker::CoreWorker(
     RAY_CHECK_OK(gcs_client_->raylet_task_table().Add(job_id, task_id, data, nullptr));
     worker_context_.SetCurrentTaskId(task_id);
   }
+
+  direct_actor_submitter_ = std::unique_ptr<CoreWorkerDirectActorTaskSubmitter>(
+      new CoreWorkerDirectActorTaskSubmitter(
+          io_service_,
+          object_interface_->CreateStoreProvider(StoreProviderType::MEMORY)));
 }
 
 CoreWorker::~CoreWorker() {
@@ -174,7 +213,7 @@ bool CoreWorker::AddActorHandle(std::unique_ptr<ActorHandle> actor_handle) {
         // submit tasks to dead actors.
       }
 
-      task_interface_->HandleDirectActorUpdate(actor_id, actor_data);
+      direct_actor_submitter_->HandleActorUpdate(actor_id, actor_data);
 
       RAY_LOG(INFO) << "received notification on actor, state="
                     << static_cast<int>(actor_data.state()) << ", actor_id: " << actor_id
@@ -192,6 +231,88 @@ ActorHandle &CoreWorker::GetActorHandle(const ActorID &actor_id) {
   auto it = actor_handles_.find(actor_id);
   RAY_CHECK(it != actor_handles_.end());
   return *it->second;
+}
+
+Status CoreWorker::SubmitTask(const RayFunction &function,
+                              const std::vector<TaskArg> &args,
+                              const TaskOptions &task_options,
+                              std::vector<ObjectID> *return_ids) {
+  TaskSpecBuilder builder;
+  const int next_task_index = worker_context_.GetNextTaskIndex();
+  const auto task_id =
+      TaskID::ForNormalTask(worker_context_.GetCurrentJobID(),
+                            worker_context_.GetCurrentTaskID(), next_task_index);
+  BuildCommonTaskSpec(builder, worker_context_.GetCurrentJobID(), task_id,
+                      worker_context_.GetCurrentTaskID(), next_task_index, GetCallerId(),
+                      function, args, task_options.num_returns, task_options.resources,
+                      {}, TaskTransportType::RAYLET, return_ids);
+  return raylet_client_->SubmitTask(builder.Build());
+}
+
+Status CoreWorker::CreateActor(const RayFunction &function,
+                               const std::vector<TaskArg> &args,
+                               const ActorCreationOptions &actor_creation_options,
+                               std::unique_ptr<ActorHandle> *actor_handle) {
+  const int next_task_index = worker_context_.GetNextTaskIndex();
+  const ActorID actor_id =
+      ActorID::Of(worker_context_.GetCurrentJobID(), worker_context_.GetCurrentTaskID(),
+                  next_task_index);
+  const TaskID actor_creation_task_id = TaskID::ForActorCreationTask(actor_id);
+  const JobID job_id = worker_context_.GetCurrentJobID();
+  std::vector<ObjectID> return_ids;
+  TaskSpecBuilder builder;
+  BuildCommonTaskSpec(
+      builder, job_id, actor_creation_task_id, worker_context_.GetCurrentTaskID(),
+      next_task_index, GetCallerId(), function, args, 1, actor_creation_options.resources,
+      actor_creation_options.placement_resources, TaskTransportType::RAYLET, &return_ids);
+  builder.SetActorCreationTaskSpec(actor_id, actor_creation_options.max_reconstructions,
+                                   actor_creation_options.dynamic_worker_options,
+                                   actor_creation_options.is_direct_call);
+
+  *actor_handle = std::unique_ptr<ActorHandle>(new ActorHandle(
+      actor_id, job_id, /*actor_cursor=*/return_ids[0], function.GetLanguage(),
+      actor_creation_options.is_direct_call, function.GetFunctionDescriptor()));
+
+  return raylet_client_->SubmitTask(builder.Build());
+}
+
+Status CoreWorker::SubmitActorTask(ActorHandle &actor_handle, const RayFunction &function,
+                                   const std::vector<TaskArg> &args,
+                                   const TaskOptions &task_options,
+                                   std::vector<ObjectID> *return_ids) {
+  // Add one for actor cursor object id for tasks.
+  const int num_returns = task_options.num_returns + 1;
+
+  const bool is_direct_call = actor_handle.IsDirectCallActor();
+  const TaskTransportType transport_type =
+      is_direct_call ? TaskTransportType::DIRECT_ACTOR : TaskTransportType::RAYLET;
+
+  // Build common task spec.
+  TaskSpecBuilder builder;
+  const int next_task_index = worker_context_.GetNextTaskIndex();
+  const TaskID actor_task_id = TaskID::ForActorTask(
+      worker_context_.GetCurrentJobID(), worker_context_.GetCurrentTaskID(),
+      next_task_index, actor_handle.GetActorID());
+  BuildCommonTaskSpec(builder, actor_handle.CreationJobID(), actor_task_id,
+                      worker_context_.GetCurrentTaskID(), next_task_index, GetCallerId(),
+                      function, args, num_returns, task_options.resources, {},
+                      transport_type, return_ids);
+
+  const ObjectID new_cursor = return_ids->back();
+  actor_handle.SetActorTaskSpec(builder, transport_type, new_cursor);
+
+  // Submit task.
+  ray::Status status;
+  if (is_direct_call) {
+    status = direct_actor_submitter_->SubmitTask(builder.Build());
+  } else {
+    status = raylet_client_->SubmitTask(builder.Build());
+  }
+
+  // Remove cursor from return ids.
+  return_ids->pop_back();
+
+  return status;
 }
 
 }  // namespace ray
